applications: serial_lte_modem: Fix UART buffer issues
1-byte long buffers are too small for asynchronous receiving on UARTE.
With such buffers, it is possible that another transfer (started by
the ENDRX_STARTRX shortcut) completes before the user handler finishes
processing data from the previous one. This results in possibility of
data being overwritten and also in invalid requests for next buffer
supplied by the driver.
This commit changes the way the application utilizes its RX buffer.
Whole buffer, instead of only one byte from it, is now supplied to the
UART driver, and to take advantage of the double buffering capability
of the driver, a second RX buffer is also introduced.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>